scene: Extrace dependency resolver
